refactor: centralize GameState type and clean up imports
Move GameState type to a dedicated file (gameStateTypes.ts) to avoid duplication and improve maintainability. Remove unnecessary comments and clean up import paths across multiple files.
